Many bug fixed, Count will show for projects & gallery

+++ reference_delete.php
@@ -1,6 +1,10 @@
 <?php
-    include("dbconnect.php");
-    $strquery=" DELETE from reference_info where ref_id = '" . $_GET['ref_id'] . "' ";
+include('dbconnect.php');
+error_reporting(0);
+include("classes/Authentication.php");
+if($isLoggedIn && $isAdmin){
+    $strquery=" DELETE from reference_info where ref_id = '". $_GET['ref_id'] ."' ";
     $results=mysql_query ($strquery);
-    header ('location: http://www.uapians.net/reference_list.php ');
+    header ("location: https://localhost/uapians/reference_list.php?CID=".$_GET['CID']);
+}
 ?>
